+++ yggdrasil/languages/C/python_wrapper.h
@@ -2,6 +2,15 @@
 extern "C" {
 #endif
 
+#ifdef YGGDRASIL_DISABLE_PYTHON_C_API
+#ifndef PyObject
+#define PyObject void
+#endif
+#ifndef npy_intp
+#define npy_intp int
+#endif
+#else // YGGDRASIL_DISABLE_PYTHON_C_API
+  
 #ifdef _DEBUG
 #undef _DEBUG
 
@@ -25,6 +34,7 @@ __declspec(dllexport) int PyObject_Print_STDOUT(PyObject* x);
 #else
 int PyObject_Print_STDOUT(PyObject* x);
 #endif
+#endif // YGGDRASIL_DISABLE_PYTHON_C_API
 
 #ifdef __cplusplus /* If this is a C++ compiler, end C linkage */
 }