+++ HISTORY.rst
@@ -21,6 +21,14 @@ TODO:
 
 * Fix bug where datatype parameters in communicators are not passed to the datatype during normalization (currently handle on python side)
 
+1.8.5 (2023-??-??) Bug fixes for Mac M1 chips (arm64) and various improvements to CLI
+------------------
+* Fix bug where yggcompile was compiling some languages multiple times
+* Allow yggcompile to be called like yggcc
+* Add --with-asan option to compilation CLI and YAML options for enabling the Clang address sanitizer
+* Allow for 128 bit types to be absent
+* Generalize option for running with a debugging tool via the --with-debugger flag for yggrun or with_debugger YAML option
+
 1.8.4 (2023-01-27) Update CI to use mamba, improve dev utilities, & various bug fixes
 ------------------
 * Update utility scripts so that requirements can be maintained in a single file with pip requirements.txt file and conda recipe generated via utils/manage_requirements.py